# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MiniPay Monitor API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface TokenBalance {
  symbol: string;
  name?: string;
  balance: string;
  usdValue: number;
  contractAddress: string;
  decimals?: number;
  /** @nullable */
  logoUrl?: string | null;
}

export interface Wallet {
  address: string;
  label: string;
  /** deposit | reward-europe | reward-us | reward-asia */
  group: string;
  totalUsdValue: number;
  tokens?: TokenBalance[];
  lastUpdated?: string;
}

export interface TokenSummary {
  symbol: string;
  totalBalance: string;
  totalUsdValue: number;
}

export interface WalletSummary {
  totalUsdValue: number;
  walletCount: number;
  depositTotal?: number;
  rewardTotal?: number;
  tokenBreakdown: TokenSummary[];
  lastUpdated?: string;
}

export interface Transaction {
  hash: string;
  from: string;
  to: string;
  value: string;
  token: string;
  /** @nullable */
  usdValue?: number | null;
  timestamp: string;
  /** in | out */
  type: string;
  /** @nullable */
  blockNumber?: number | null;
}

export interface TransactionPage {
  transactions: Transaction[];
  total: number;
  page: number;
  limit: number;
}

export interface Alert {
  id: number;
  walletAddress: string;
  walletLabel: string;
  token: string;
  /** increase | decrease */
  changeType: string;
  oldValue: string;
  newValue: string;
  /** @nullable */
  changePercent?: number | null;
  isRead: boolean;
  createdAt: string;
}

export interface UnreadCount {
  count: number;
}