# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: MiniPay Monitor API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /wallets/{address}/balances:
    get:
      operationId: getWalletBalances
      tags: [wallets]
      summary: Get token balances for a specific wallet
      parameters:
        - name: address
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Token balances
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TokenBalance"
...
  /alerts/{id}/read:
    patch:
      operationId: markAlertRead
      tags: [alerts]
      summary: Mark an alert as read
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Alert marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
components:
  schemas:
...
    TokenBalance:
      type: object
      required: [symbol, balance, usdValue, contractAddress]
      properties:
        symbol:
          type: string
        name:
          type: string
        balance:
          type: string
        usdValue:
          type: number
        contractAddress:
          type: string
        decimals:
          type: integer
        logoUrl:
          type: ["string", "null"]
...
    Alert:
      type: object
      required: [id, walletAddress, walletLabel, token, changeType, oldValue, newValue, isRead, createdAt]
      properties:
        id:
          type: integer
        walletAddress:
          type: string
        walletLabel:
          type: string
        token:
          type: string
        changeType:
          type: string
          description: "increase | decrease"
        oldValue:
          type: string
        newValue:
          type: string
        changePercent:
          type: ["number", "null"]
        isRead:
          type: boolean
        createdAt:
          type: string
...