# lib/api-zod/src/generated/types/alert.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MiniPay Monitor API
 * OpenAPI spec version: 0.1.0
 */

export interface Alert {
  id: number;
  walletAddress: string;
  walletLabel: string;
  token: string;
  /** increase | decrease */
  changeType: string;
  oldValue: string;
  newValue: string;
  /** @nullable */
  changePercent?: number | null;
  isRead: boolean;
  createdAt: string;
}