# lib/api-zod/src/generated/types/tokenBalance.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * MiniPay Monitor API
 * OpenAPI spec version: 0.1.0
 */

export interface TokenBalance {
  symbol: string;
  name?: string;
  balance: string;
  usdValue: number;
  contractAddress: string;
  decimals?: number;
  /** @nullable */
  logoUrl?: string | null;
}